Move _legacy.cpp contents to UI project
