+++ src/openrct2/libopenrct2.vcxproj
@@ -1083,7 +1083,6 @@
     <ClCompile Include="util\Util.cpp" />
     <ClCompile Include="Version.cpp" />
     <ClCompile Include="windows\Intent.cpp" />
-    <ClCompile Include="windows\_legacy.cpp" />
     <ClCompile Include="world\tile_element\EntranceElement.cpp" />
     <ClCompile Include="world\Banner.cpp" />
     <ClCompile Include="world\Climate.cpp" />
