+++ src/openrct2/windows/Intent.h
@@ -52,6 +52,7 @@ enum IntentAction
     INTENT_ACTION_REMOVE_PROVISIONAL_ELEMENTS,
     INTENT_ACTION_RESTORE_PROVISIONAL_ELEMENTS,
     INTENT_ACTION_REMOVE_PROVISIONAL_FOOTPATH,
+    INTENT_ACTION_REMOVE_PROVISIONAL_TRACK_PIECE,
 
     INTENT_ACTION_NULL = 255,
 };
